+++ ethergo/submitter/doc.go
@@ -1,0 +1,2 @@
+// Package submitter provides a simple interface for submitting txs to multiple chains and retrying.
+package submitter