+++ packages/sdk-router/src/sdk.ts
@@ -3,7 +3,7 @@ import invariant from 'tiny-invariant'
 import { BigNumber } from '@ethersproject/bignumber'
 import { BytesLike } from '@ethersproject/bytes'
 import { PopulatedTransaction } from 'ethers'
-import { AddressZero } from '@ethersproject/constants'
+import { AddressZero, Zero } from '@ethersproject/constants'
 import { Interface } from '@ethersproject/abi'
 import { Contract } from '@ethersproject/contracts'
 
@@ -293,19 +293,79 @@ class SynapseSDK {
   public async calculateAddLiquidity(
     chainId: number,
     poolAddress: string,
-    amounts: BigNumber[]
-  ): Promise<BigNumber> {
+    amounts: Record<string, BigNumber>
+  ): Promise<{ amount: BigNumber; routerAddress: string }> {
     const router: SynapseRouter = this.synapseRouters[chainId]
-    return router.routerContract.calculateAddLiquidity(poolAddress, amounts)
+    const poolTokens = await router.routerContract.poolTokens(poolAddress)
+    const amountArr: BigNumber[] = []
+    poolTokens.map((token) => {
+      amountArr.push(amounts[token.token] ?? Zero)
+    })
+    if (amountArr.filter((amount) => !amount.isZero()).length === 0) {
+      return { amount: Zero, routerAddress: router.routerContract.address }
+    }
+    return {
+      amount: await router.routerContract.calculateAddLiquidity(
+        poolAddress,
+        amountArr
+      ),
+      routerAddress: router.routerContract.address,
+    }
   }
 
   public async calculateRemoveLiquidity(
     chainId: number,
     poolAddress: string,
     amount: BigNumber
-  ): Promise<BigNumber[]> {
+  ): Promise<{
+    amounts: Record<string, { value: BigNumber; index: number }>
+    routerAddress: string
+  }> {
     const router: SynapseRouter = this.synapseRouters[chainId]
-    return router.routerContract.calculateRemoveLiquidity(poolAddress, amount)
+    const amounts = await router.routerContract.calculateRemoveLiquidity(
+      poolAddress,
+      amount
+    )
+    const poolTokens = await router.routerContract.poolTokens(poolAddress)
+    const amountsOut: Record<string, { value: BigNumber; index: number }> = {}
+    poolTokens.map((token, index) => {
+      amountsOut[token.token] = { value: amounts[index], index }
+    })
+    return {
+      amounts: amountsOut,
+      routerAddress: router.routerContract.address,
+    }
+  }
+
+  public async calculateRemoveLiquidityOne(
+    chainId: number,
+    poolAddress: string,
+    amount: BigNumber,
+    token: string
+  ): Promise<{
+    amount: { value: BigNumber; index: number }
+    routerAddress: string
+  }> {
+    const router: SynapseRouter = this.synapseRouters[chainId]
+
+    let poolIndex = 0
+    const poolTokens = await router.routerContract.poolTokens(poolAddress)
+    poolTokens.map((poolToken, index) => {
+      if (poolToken.token === token) {
+        poolIndex = index
+      }
+    })
+
+    const outAmount = await router.routerContract.calculateWithdrawOneToken(
+      poolAddress,
+      amount,
+      poolIndex
+    )
+
+    return {
+      amount: { value: outAmount, index: poolIndex },
+      routerAddress: router.routerContract.address,
+    }
   }
 }