+++ packages/synapse-interface/utils/approveToken.ts
@@ -2,21 +2,20 @@ import { erc20ABI } from 'wagmi'
 import { fetchSigner } from '@wagmi/core'
 import { Contract } from 'ethers'
 import { MaxInt256 } from '@ethersproject/constants'
-
-export const approveToken = async (address, chainId, fromTokenAddress) => {
+import { BigNumber } from '@ethersproject/bignumber'
+export const approveToken = async (
+  address: string,
+  chainId: number,
+  tokenAddress: string,
+  amount?: BigNumber
+) => {
   // TODO store this erc20 and signer retrieval in a state in a parent component
-  const wallet = await fetchSigner({
+  const signer = await fetchSigner({
     chainId,
   })
 
-  console.log(
-    'address, chainId, fromTokenAddress',
-    address,
-    chainId,
-    fromTokenAddress
-  )
-  const erc20 = new Contract(fromTokenAddress, erc20ABI, wallet)
-  const approveTx = await erc20.approve(address, MaxInt256)
+  const erc20 = new Contract(tokenAddress, erc20ABI, signer)
+  const approveTx = await erc20.approve(address, amount ?? MaxInt256)
 
   try {
     await approveTx.wait()