+++ services/omnirpc/client/client.go
@@ -2,13 +2,17 @@ package client
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"github.com/synapsecns/sanguine/core/metrics"
 	"github.com/synapsecns/sanguine/ethergo/client"
+	"github.com/synapsecns/sanguine/ethergo/submitter"
+	"math/big"
 )
 
 // RPCClient is an interface for the omnirpc service.
 type RPCClient interface {
+	submitter.ClientFetcher
 	// GetEndpoint returns the endpoint for the given chainID and confirmations.
 	GetEndpoint(chainID, confirmations int) string
 	// GetDefaultEndpoint returns the endpoint with the default confirmation count for the chain id.
@@ -23,6 +27,7 @@ type rpcClient struct {
 	config   *rpcOptions
 	endpoint string
 	handler  metrics.Handler
+	opts     []client.Options
 }
 
 // NewOmnirpcClient creates a new RPCClient.
@@ -31,8 +36,17 @@ func NewOmnirpcClient(endpoint string, handler metrics.Handler, options ...Optio
 	c.config = makeOptions(options)
 	c.endpoint = endpoint
 	c.handler = handler
+	c.opts = append(c.opts, client.Capture(c.config.captureReqRes))
 
 	return &c
+}
+
+func (c *rpcClient) GetClient(ctx context.Context, chainID *big.Int) (client.EVM, error) {
+	if !chainID.IsInt64() {
+		return nil, errors.New("chain id is not a uint64")
+	}
+
+	return c.GetChainClient(ctx, int(chainID.Uint64()))
 }
 
 func (c *rpcClient) GetEndpoint(chainID, confirmations int) string {
@@ -48,7 +62,7 @@ func (c *rpcClient) GetDefaultEndpoint(chainID int) string {
 
 func (c *rpcClient) GetConfirmationsClient(ctx context.Context, chainID, confirmations int) (client.EVM, error) {
 	endpoint := c.GetEndpoint(chainID, confirmations)
-	chainClient, err := client.DialBackend(ctx, endpoint, c.handler)
+	chainClient, err := client.DialBackend(ctx, endpoint, c.handler, c.opts...)
 	if err != nil {
 		return nil, fmt.Errorf("could not dial backend: %w", err)
 	}
@@ -57,7 +71,7 @@ func (c *rpcClient) GetConfirmationsClient(ctx context.Context, chainID, confirm
 
 func (c *rpcClient) GetChainClient(ctx context.Context, chainID int) (client.EVM, error) {
 	endpoint := c.GetDefaultEndpoint(chainID)
-	chainClient, err := client.DialBackend(ctx, endpoint, c.handler)
+	chainClient, err := client.DialBackend(ctx, endpoint, c.handler, c.opts...)
 	if err != nil {
 		return nil, fmt.Errorf("could not dial backend: %w", err)
 	}