+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.21](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.1.20...@synapsecns/sdk-router@0.1.21) (2023-05-03)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.1.20](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.1.19...@synapsecns/sdk-router@0.1.20) (2023-05-02)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.1.19](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.1.18...@synapsecns/sdk-router@0.1.19) (2023-05-02)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.1.18](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.1.17...@synapsecns/sdk-router@0.1.18) (2023-04-20)
 
 **Note:** Version bump only for package @synapsecns/sdk-router